Rename enums.go to constants.go
Moved any constants from enums.go to an existing constants.go.
