+++ sdk/storage/azblob/service/constants.go
@@ -6,6 +6,8 @@
 
 package service
 
+import "github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/internal/generated"
+
 const (
 	// ContainerNameRoot is the special Azure Storage name used to identify a storage account's root container.
 	ContainerNameRoot = "$root"
@@ -13,3 +15,83 @@ const (
 	// ContainerNameLogs is the special Azure Storage name used to identify a storage account's logs container.
 	ContainerNameLogs = "$logs"
 )
+
+// SKUName defines values for SkuName - LRS, GRS, RAGRS, ZRS, Premium LRS
+type SKUName string
+
+const (
+	SKUNameStandardLRS   SKUName = "Standard_LRS"
+	SKUNameStandardGRS   SKUName = "Standard_GRS"
+	SKUNameStandardRAGRS SKUName = "Standard_RAGRS"
+	SKUNameStandardZRS   SKUName = "Standard_ZRS"
+	SKUNamePremiumLRS    SKUName = "Premium_LRS"
+)
+
+// PossibleSKUNameValues returns the possible values for the SKUName const type.
+func PossibleSKUNameValues() []SKUName {
+	return []SKUName{
+		SKUNameStandardLRS,
+		SKUNameStandardGRS,
+		SKUNameStandardRAGRS,
+		SKUNameStandardZRS,
+		SKUNamePremiumLRS,
+	}
+}
+
+// ListContainersIncludeType defines values for ListContainersIncludeType
+type ListContainersIncludeType string
+
+const (
+	ListContainersIncludeTypeMetadata ListContainersIncludeType = "metadata"
+	ListContainersIncludeTypeDeleted  ListContainersIncludeType = "deleted"
+	ListContainersIncludeTypeSystem   ListContainersIncludeType = "system"
+)
+
+// PossibleListContainersIncludeTypeValues returns the possible values for the ListContainersIncludeType const type.
+func PossibleListContainersIncludeTypeValues() []ListContainersIncludeType {
+	return []ListContainersIncludeType{
+		ListContainersIncludeTypeMetadata,
+		ListContainersIncludeTypeDeleted,
+		ListContainersIncludeTypeSystem,
+	}
+}
+
+// AccountKind defines values for AccountKind
+type AccountKind = generated.AccountKind
+
+const (
+	AccountKindStorage          = generated.AccountKindStorage
+	AccountKindBlobStorage      = generated.AccountKindBlobStorage
+	AccountKindStorageV2        = generated.AccountKindStorageV2
+	AccountKindFileStorage      = generated.AccountKindFileStorage
+	AccountKindBlockBlobStorage = generated.AccountKindBlockBlobStorage
+)
+
+// PossibleAccountKindValues returns the possible values for the AccountKind const type.
+func PossibleAccountKindValues() []AccountKind {
+	return []AccountKind{
+		AccountKindStorage,
+		AccountKindBlobStorage,
+		AccountKindStorageV2,
+		AccountKindFileStorage,
+		AccountKindBlockBlobStorage,
+	}
+}
+
+// BlobGeoReplicationStatus - The status of the secondary location
+type BlobGeoReplicationStatus = generated.BlobGeoReplicationStatus
+
+const (
+	BlobGeoReplicationStatusLive        = generated.BlobGeoReplicationStatusLive
+	BlobGeoReplicationStatusBootstrap   = generated.BlobGeoReplicationStatusBootstrap
+	BlobGeoReplicationStatusUnavailable = generated.BlobGeoReplicationStatusUnavailable
+)
+
+// PossibleBlobGeoReplicationStatusValues returns the possible values for the BlobGeoReplicationStatus const type.
+func PossibleBlobGeoReplicationStatusValues() []BlobGeoReplicationStatus {
+	return []BlobGeoReplicationStatus{
+		BlobGeoReplicationStatusLive,
+		BlobGeoReplicationStatusBootstrap,
+		BlobGeoReplicationStatusUnavailable,
+	}
+}
